matrix-app-front-2/matrix-app-front/src/core/modules/auth/enrollment/infrastructure/factory/enrollment.factory.ts
import TypeProviderEnum from 'src/core/enums/type-provider.enum';
import { IEnrollmentRepository } from '../../repository/enrollment.repository';
import EnrollmentProvider from '../driver-adapter/enrollment.provider';
import EnrollmentProviderMock from '../driver-adapter-mock/enrollment.provider.mock';

class EnrollmentFactory {
  static getInstance(typeProvider: string = TypeProviderEnum.PROVIDER): IEnrollmentRepository {
    switch (typeProvider) {
      case TypeProviderEnum.PROVIDER:
        return new EnrollmentProvider();
      case TypeProviderEnum.MOCK:
        return new EnrollmentProviderMock();
      default:
        return new EnrollmentProvider();
    }
  }
}

export default EnrollmentFactory;
